Updated to Minecraft 1.20
If you notice any errors, please report them directly to me.

Bugfixes:

+ Fixed a Problem with the WaterLevel.

If the config option World.WaterHeight was used before, this update will change the look of the new chunks. Since the old version is still compatible, I do not recommend an update for this particular case.

Added:

+ Support for Spigot 1.20 (Version 1.3.0 has already been compatible with Minecraft 1.20)
+ The Config.yml Option "World.Depth" to set the depth of the World (Default -64 for Minecraft > 1.18, old configs are set to the 0)
+ The Congif.yml Option "World.GenerateWater" to generate water within the ALF World.
The Config.yml option "World.OresChance" to change the number of ores to be generated. (Default 100%)

[​IMG]

Removed:

+ The Config.yml Option "World.GenerateOres".
+ The Config.yml Option "World.WaterHeight".